requirements.txt
ffmpeg-python
ttkbootstrap
numpy